Split orchestration store tests for lint policy
